week07_application_deployment/homework/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
prometheus_client==0.15.0
torch==2.0.0
torchvision==0.15.0
requests==2.28.2
protobuf==4.23.3
grpcio==1.54.0
grpcio-tools==1.54.0
supervisor==4.2.5
Pillow==9.5.0
numpy==1.25.2
pytest
furl